requirements.txt
absl-py==2.1.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cvlib==0.2.7
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
executing==2.2.0
ffmpeg-python==0.2.0
Flask==3.0.3
flask-cors==5.0.1
flask-ngrok==0.0.25
flatbuffers==24.3.25
fonttools==4.53.1
future==1.0.0
gast==0.6.0
geocoder==1.38.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.67.1
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imbalanced-learn==0.12.4
imutils==0.5.4
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==6.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.13.2
nest-asyncio==1.6.0
numpy==2.0.1
openai==1.97.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
progressbar==2.5
prompt_toolkit==3.0.50
protobuf==5.28.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==18.0.0
PyAudio==0.2.14
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2024.2
pywin32==310
pyzmq==26.3.0
ratelim==0.1.6
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.14.1
stack-data==0.6.3
streamlit==1.40.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.1.1
wheel==0.44.0
wrapt==1.16.0
